refactor: app styles to app/styles
