+++ src/app/components/option-item/option-item.tsx
@@ -1,4 +1,4 @@
-import { ArrowIcon } from '@/icons';
+import { ArrowIcon } from '@/shared/icons';
 import { LinkCustom } from '@/shared/ui/link-custom';
 
 import './option-item.scss';
