+++ src/app/styles/index.scss
@@ -1,5 +1,3 @@
-@import './normalize';
-
 html {
   scroll-behavior: smooth;
 }
@@ -38,9 +36,7 @@ figure {
   display: flex;
   align-items: center;
   justify-content: center;
-
   width: 100%;
-
   text-align: left;
 }
 
@@ -55,7 +51,7 @@ figure {
 
   width: 100%;
   max-width: 1440px;
-  padding: 80px 120px 40px 120px;
+  padding: 80px 120px 40px;
 }
 
 .column-2 {
