+++ src/features/certification/certification.tsx
@@ -1,5 +1,5 @@
 import classNames from 'classnames/bind';
-import { CourseNames } from '@/data/communication.data';
+import { CourseNames } from '@/shared/data/communication.data';
 import { Paragraph } from '@/shared/ui/paragraph';
 import { Title } from '@/shared/ui/title';
