+++ src/features/course-main/course-main.tsx
@@ -4,9 +4,9 @@ import { DateLang } from '@/app/components';
 
 import { Course } from '@/app/types';
 import Image from '@/features/image';
-import { ArrowIcon } from '@/icons';
 import { selectCourse } from '@/shared/hooks/use-course-by-title/utils/select-course';
 import { useTitle } from '@/shared/hooks/use-title';
+import { ArrowIcon } from '@/shared/icons';
 import { LinkCustom } from '@/shared/ui/link-custom';
 import { SectionLabel } from '@/shared/ui/section-label';
 import { Subtitle } from '@/shared/ui/subtitle';
