+++ src/features/home/community/community.tsx
@@ -1,6 +1,6 @@
 import Image from '@/features/image';
-import { FacebookIcon, InstagramIcon, LinkedInIcon, TelegramIcon, YouTubeIcon } from '@/icons';
 import image from '@/shared/assets/welcome.webp';
+import { FacebookIcon, InstagramIcon, LinkedInIcon, TelegramIcon, YouTubeIcon } from '@/shared/icons';
 import { SocialMedia, SocialMediaProps } from '@/shared/ui/social-media';
 import { Subtitle } from '@/shared/ui/subtitle';
 import { Title, TitleType } from '@/shared/ui/title';
