+++ src/features/home/merch/merch.tsx
@@ -1,7 +1,7 @@
 import { LINKS } from '@/app/const';
 import Image from '@/features/image';
-import { ArrowIcon } from '@/icons';
 import image from '@/shared/assets/merch.webp';
+import { ArrowIcon } from '@/shared/icons';
 import { LinkCustom } from '@/shared/ui/link-custom';
 import { Paragraph } from '@/shared/ui/paragraph';
 import { SectionLabel } from '@/shared/ui/section-label';
