+++ src/features/home/pictures/pictures.tsx
@@ -1,7 +1,6 @@
 // requires a loader
 import { Carousel } from 'react-responsive-carousel';
 import Image from '@/features/image';
-import { FacebookIcon, InstagramIcon } from '@/icons';
 import photo1 from '@/shared/assets/photo-1.webp';
 import photo2 from '@/shared/assets/photo-2.webp';
 import photo3 from '@/shared/assets/photo-3.webp';
@@ -11,6 +10,7 @@ import photo6 from '@/shared/assets/photo-6.webp';
 import photo7 from '@/shared/assets/photo-7.webp';
 import photo8 from '@/shared/assets/photo-8.webp';
 import photo9 from '@/shared/assets/photo-9.webp';
+import { FacebookIcon, InstagramIcon } from '@/shared/icons';
 import { Paragraph } from '@/shared/ui/paragraph';
 import { SocialMedia, SocialMediaProps } from '@/shared/ui/social-media';
 import { Title } from '@/shared/ui/title';
