+++ src/features/home/speakers/speakers.tsx
@@ -1,6 +1,6 @@
 import Image from '@/features/image';
-import { EmailIcon } from '@/icons';
 import image from '@/shared/assets/speakers-wanted.webp';
+import { EmailIcon } from '@/shared/icons';
 import { Subtitle } from '@/shared/ui/subtitle';
 import { Title, TitleType } from '@/shared/ui/title';
