+++ src/features/school/courses/ui/CourseCard.tsx
@@ -1,5 +1,5 @@
 import { Course } from '@/app/types';
-import { ArrowIcon } from '@/icons';
+import { ArrowIcon } from '@/shared/icons';
 import { LinkCustom } from '@/shared/ui/link-custom';
 
 type addFields = {
