+++ src/main.tsx
@@ -3,7 +3,7 @@ import ReactDOM from 'react-dom/client';
 import App from './app/app';
 import '@/app/services/dayjs';
 
-import '@/styles/index.scss';
+import '@/app/styles/index.scss';
 
 ReactDOM.createRoot(document.getElementById('root')!).render(
   <React.StrictMode>
