+++ src/shared/data/communication.data.ts
