+++ src/shared/data/courseTitles.data.ts
