+++ src/shared/icons/angular-icon.tsx
