+++ src/shared/icons/aws.tsx
