+++ src/shared/icons/discord-logo.tsx
