+++ src/shared/icons/email.tsx
