+++ src/shared/icons/epam.tsx
