+++ src/shared/icons/javascript-icon.tsx
