+++ src/shared/icons/logo.tsx
