+++ src/shared/icons/open-source-philosophy-icon.tsx
