+++ src/shared/icons/open-to-everyone-icon.tsx
