+++ src/shared/icons/react-icon.tsx
