+++ src/shared/icons/rs-banner.tsx
