+++ src/shared/icons/text-link-icon.tsx
