+++ vite.config.ts
@@ -20,9 +20,10 @@ export default defineConfig({
     preprocessorOptions: {
       scss: {
         additionalData: `
-          @import "./src/styles/_constants.scss";
-          @import "./src/styles/_mixins.scss";
-          @import "./src/styles/_placeholders.scss";
+          @import "./src/app/styles/_constants.scss";
+          @import "./src/app/styles/_mixins.scss";
+          @import "./src/app/styles/_placeholders.scss";
+          @import "./src/app/styles/_normalize.scss";
         `,
       },
     },
